Keep dependency modifiers in one module
Keep the public modifier methods and their private lookup index together so the implementation can be read in one place. Remove the child module and its extra visibility declarations.
